requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
openpyxl==3.1.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
bcrypt==4.1.2
PyJWT==2.8.1
SQLAlchemy==2.0.23
passlib==1.7.4